ENH: Move rtkStoreSparseMatrixSplatWeightMultiplication to include
